sfdp_services: support larger NAT pools
The old code only had a TODO and an assert.

+ Actually use the "remaining" vector of addresses when needed.
+ Use u32 for size of pool.
+ Convert src IP into machine order when computing pool index.
- Do not fix the off-by-one error in vec_validate yet,
  later commit adds unit tests for that.

Type: fix
Fixes: 28dbf5aee6bfd783e7336d561a1fe03f455c7e9b

+++ src/plugins/sfdp_services/base/nat/nat.h
@@ -67,7 +67,7 @@ typedef struct
 {
   CLIB_CACHE_LINE_ALIGN_MARK (cache0);
   u16 flags;
-  u16 num;
+  u32 num;
   ip4_address_t addr[NAT_ALLOC_POOL_ARRAY_SZ];
   ip4_address_t *remaining;
 } nat_alloc_pool_t;
